+++ Pharmacie-project/Api/Models/User.cs
@@ -19,7 +19,7 @@ public class User
     public UserRole Role { get; set; }
 
     public Guid? PharmacyId { get; set; }
-    public decimal CostPerKM { get; set; }
+    public int CostPerKM { get; set; }
     public bool valider { get; set; }
     public  DateTime? verifiedAt { get; set; }
